Agents, buildings, and items support "soft" collisions.
Soft collisions offer a way to detect if a tile is occupied by an entity but
still pathable through. Useful when we don't want to place buildings on top of
items or people or trees.

One thing to note is that soft collisions are backed by a semaphore. This is
because since multiple entities can occupy a single tile, and because we need
to mark tiles as "softCollison === false" when all entities are gone, we need
to keep track of how many are in the tile at a given time.